+++ crates/full-node/sov-metrics/src/lib.rs
@@ -3,11 +3,11 @@
 
 mod influxdb;
 
-#[cfg(feature = "gas-constant-estimation")]
-pub use influxdb::GAS_CONSTANTS;
 pub use influxdb::{
     init_metrics_tracker, timestamp, track_metrics, BatchMetrics, BatchOutcome, HttpMetrics,
     MetricsTracker, MonitoringConfig, RunnerMetrics, SlotProcessingMetrics, SovRollupMetrics,
     TransactionEffect, TransactionProcessingMetrics, UserSpaceSlotProcessingMetrics, ZkCircuit,
     ZkProvingTime, ZkVmExecutionChunk,
 };
+#[cfg(feature = "gas-constant-estimation")]
+pub use influxdb::{GasConstantTracker, GAS_CONSTANTS};